Split variables in expenses into separate files
