+++ openfisca_us/variables/expense/housing_cost.py
@@ -1,0 +1,10 @@
+from openfisca_us.model_api import *
+
+
+class housing_cost(Variable):
+    value_type = float
+    entity = SPMUnit
+    label = "Housing cost"
+    documentation = "Housing cost for this SPM unit"
+    unit = "currency-USD"
+    definition_period = YEAR
