+++ openfisca_us/variables/expense/phone_cost.py
@@ -1,0 +1,10 @@
+from openfisca_us.model_api import *
+
+
+class phone_cost(Variable):
+    value_type = float
+    entity = SPMUnit
+    label = "Phone cost"
+    documentation = "Phone line cost for this SPM unit"
+    unit = "currency-USD"
+    definition_period = YEAR
