+++ openfisca_us/variables/expense/spm_unit_federal_tax.py
@@ -1,0 +1,9 @@
+from openfisca_us.model_api import *
+
+
+class spm_unit_federal_tax(Variable):
+    value_type = float
+    entity = SPMUnit
+    label = u"SPM unit federal tax"
+    definition_period = YEAR
+    unit = "currency-USD"
