+++ openfisca_us/variables/expense/spm_unit_medical_expenses.py
@@ -1,0 +1,9 @@
+from openfisca_us.model_api import *
+
+
+class spm_unit_medical_expenses(Variable):
+    value_type = float
+    entity = SPMUnit
+    label = u"SPM unit medical expenses"
+    definition_period = YEAR
+    unit = "currency-USD"
