+++ openfisca_us/variables/expense/spm_unit_state_tax.py
@@ -1,0 +1,9 @@
+from openfisca_us.model_api import *
+
+
+class spm_unit_state_tax(Variable):
+    value_type = float
+    entity = SPMUnit
+    label = u"SPM unit state tax"
+    definition_period = YEAR
+    unit = "currency-USD"
